refactor: move template backend tests to test_templates.py
